add movie list activity, associated views and presenter

+++ app/src/main/java/com/d3m0li5h3r/samples/moviesbrowser/api/models/DiscoverMovieResponse.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
 
 data class DiscoverMovieResponse(
-        @Expose @SerializedName(value = "page") val currentPage: Long,
+        @Expose @SerializedName(value = "page") val currentPage: Int,
         @Expose @SerializedName(value = "total_results") val totalResults: Long,
-        @Expose @SerializedName(value = "total_pages") val totalPages: Long,
+        @Expose @SerializedName(value = "total_pages") val totalPages: Int,
         @Expose @SerializedName(value = "results") val movies: List<Movie>
 )
